rename packakge to nn.innit
