+++ docs/src/python/nn.rst
@@ -180,4 +180,4 @@ In detail:
    nn/layers
    nn/functions
    nn/losses
-   nn/initializers
+   nn/init
