+++ python/mlx/nn/__init__.py
@@ -1,5 +1,5 @@
 # Copyright © 2023 Apple Inc.
 
-from mlx.nn import initializers, losses
+from mlx.nn import init, losses
 from mlx.nn.layers import *
 from mlx.nn.utils import value_and_grad
